Split api routes with proper namings, remove tests from pkg folder, put build package into internal, remove unused code, refactor some code
