+++ api/staking.go
@@ -4,10 +4,12 @@ import (
 	"github.com/gin-gonic/gin"
 	"github.com/trustwallet/blockatlas/coin"
 	"github.com/trustwallet/blockatlas/pkg/blockatlas"
+	"github.com/trustwallet/blockatlas/pkg/errors"
 	"github.com/trustwallet/blockatlas/pkg/ginutils"
 	"github.com/trustwallet/blockatlas/pkg/ginutils/gincache"
+	"github.com/trustwallet/blockatlas/pkg/logger"
 	"github.com/trustwallet/blockatlas/platform"
-	"strconv"
+	services "github.com/trustwallet/blockatlas/services/assets"
 	"time"
 )
 
@@ -99,81 +101,90 @@ func makeStakingDelegationsSimpleBatchRoute(router gin.IRouter) {
 	}))
 }
 
-// @Description Get collection categories
-// @ID collection_categories_v2
-// @Summary Get list of collections from a specific coin and addresses
+// @Summary Get Validators
+// @ID validators
+// @Description Get validators from the address
 // @Accept json
 // @Produce json
-// @Tags Collections
-// @Param data body string true "Payload" default({"60": ["0xb3624367b1ab37daef42e1a3a2ced012359659b0"]})
+// @Tags Staking
+// @Param coin path string true "the coin name" default(cosmos)
 // @Success 200 {object} blockatlas.DocsResponse
-// @Router /v2/collectibles/categories [post]
-//TODO: remove once most of the clients will be updated (deadline: March 17th)
-func oldMakeCategoriesBatchRoute(router gin.IRouter) {
-	router.POST("/collectibles/categories", func(c *gin.Context) {
-		var reqs map[string][]string
-		if err := c.BindJSON(&reqs); err != nil {
+// @Failure 500 {object} ginutils.ApiError
+// @Router /v2/{coin}/staking/validators [get]
+func makeStakingValidatorsRoute(router gin.IRouter, api blockatlas.Platform) {
+	var stakingAPI blockatlas.StakeAPI
+	stakingAPI, _ = api.(blockatlas.StakeAPI)
+
+	if stakingAPI == nil {
+		return
+	}
+
+	router.GET("/staking/validators", gincache.CacheMiddleware(time.Hour, func(c *gin.Context) {
+		results, err := services.GetActiveValidators(stakingAPI)
+		if err != nil {
+			logger.Error(err)
+			ginutils.ErrorResponse(c).Message(err.Error()).Render()
+			return
+		}
+		ginutils.RenderSuccess(c, blockatlas.DocsResponse{Docs: results})
+	}))
+}
+
+// @Summary Get Stake Delegations
+// @ID delegations
+// @Description Get stake delegations from the address
+// @Accept json
+// @Produce json
+// @Tags Staking
+// @Param coin path string true "the coin name" default(tron)
+// @Param address path string true "the query address" default(TPJYCz8ppZNyvw7pTwmjajcx4Kk1MmEUhD)
+// @Success 200 {object} blockatlas.DelegationResponse
+// @Failure 500 {object} ginutils.ApiError
+// @Router /v2/{coin}/staking/delegations/{address} [get]
+func makeStakingDelegationsRoute(router gin.IRouter, api blockatlas.Platform) {
+	var stakingAPI blockatlas.StakeAPI
+	stakingAPI, _ = api.(blockatlas.StakeAPI)
+
+	if stakingAPI == nil {
+		return
+	}
+
+	router.GET("/staking/delegations/:address", func(c *gin.Context) {
+		response, err := getDelegationResponse(stakingAPI, c.Param("address"))
+		if err != nil {
 			ginutils.ErrorResponse(c).Message(err.Error()).Render()
 			return
 		}
 
-		batch := make(blockatlas.CollectionPage, 0)
-		for key, addresses := range reqs {
-			coinId, err := strconv.Atoi(key)
-			if err != nil {
-				continue
-			}
-			p, ok := platform.CollectionAPIs[uint(coinId)]
-			if !ok {
-				continue
-			}
-			for _, address := range addresses {
-				collections, err := p.OldGetCollections(address)
-				if err != nil {
-					continue
-				}
-				batch = append(batch, collections...)
-			}
-		}
-		ginutils.RenderSuccess(c, batch)
+		ginutils.RenderSuccess(c, response)
 	})
 }
 
-// @Description Get collection categories
-// @ID collection_categories_v3
-// @Summary Get list of collections from a specific coin and addresses
-// @Accept json
-// @Produce json
-// @Tags Collections
-// @Param data body string true "Payload" default({"60": ["0xb3624367b1ab37daef42e1a3a2ced012359659b0"]})
-// @Success 200 {object} blockatlas.DocsResponse
-// @Router /v3/collectibles/categories [post]
-func makeCategoriesBatchRoute(router gin.IRouter) {
-	router.POST("/collectibles/categories", func(c *gin.Context) {
-		var reqs map[string][]string
-		if err := c.BindJSON(&reqs); err != nil {
-			ginutils.ErrorResponse(c).Message(err.Error()).Render()
-			return
-		}
+func getDelegationResponse(p blockatlas.StakeAPI, address string) (blockatlas.DelegationResponse, error) {
+	delegations, err := p.GetDelegations(address)
+	if err != nil {
+		return blockatlas.DelegationResponse{
+			StakingResponse: getStakingResponse(p),
+		}, errors.E("Unable to fetch delegations list", err)
+	}
+	balance, err := p.UndelegatedBalance(address)
+	if err != nil {
+		return blockatlas.DelegationResponse{
+			StakingResponse: getStakingResponse(p),
+		}, errors.E("Unable to fetch undelegated balance", err)
+	}
+	return blockatlas.DelegationResponse{
+		Balance:         balance,
+		Delegations:     delegations,
+		Address:         address,
+		StakingResponse: getStakingResponse(p),
+	}, nil
+}
 
-		batch := make(blockatlas.CollectionPage, 0)
-		for key, addresses := range reqs {
-			coinId, err := strconv.Atoi(key)
-			if err != nil {
-				continue
-			}
-			p, ok := platform.CollectionAPIs[uint(coinId)]
-			if !ok {
-				continue
-			}
-			for _, address := range addresses {
-				collections, err := p.GetCollections(address)
-				if err != nil {
-					continue
-				}
-				batch = append(batch, collections...)
-			}
-		}
-		ginutils.RenderSuccess(c, batch)
-	})
+func getStakingResponse(p blockatlas.StakeAPI) blockatlas.StakingResponse {
+	c := p.Coin()
+	return blockatlas.StakingResponse{
+		Coin:    c.External(),
+		Details: p.GetDetails(),
+	}
 }
