+++ tests/functional/testdata/coin_fixtures.json
